Put missing functionality into place in item and file system managers
