+++ openpassword/agile_keychain/_item_manager.py
@@ -1,5 +1,6 @@
 import os
 import json
+import glob
 
 from openpassword.agile_keychain.agile_keychain_item import EncryptedItem
 
@@ -16,21 +17,17 @@ class ItemManager:
 
         return EncryptedItem(data)
 
+    def get_all_items(self):
+        item_paths = glob.glob(os.path.join(self._base_path, "data", "default", "*.1password"))
+
+        items = []
+        for item_path in item_paths:
+            basename = os.path.basename(item_path)
+            item_id, _ = os.path.splitext(basename)
+            items.append(self.get_by_id(item_id))
+
+        return items
+
     def save_item(self, item):
         with open(os.path.join(self._base_path, "data", "default", "{0}.1password".format(item['uuid'])), "w") as file:
             json.dump(item, file)
-
-        # key = self._get_key_for_item(data)
-
-        # encrypted = b64decode(data['encrypted'])
-        # init_vector = encrypted[8:16]
-        # derived_key = Crypto.derive_key(key.decrypted_key, init_vector)
-        # decrypted = Crypto.decrypt(derived_key[0:16], derived_key[16:], encrypted[16:])
-        # decrypted = strip_byte_padding(decrypted)
-        # decrypted_data = json.loads(decrypted.decode('ascii'))
-
-        # item = AgileKeychainItem()
-        # item.id = data['uuid']
-        # item.secrets = decrypted_data
-
-        # return item
