+++ tests/integration/openpassword/agile_keychain/test_item_manager.py
@@ -20,6 +20,27 @@ class ItemManagerTest:
 
         assert item['uuid'] == "5F7210FD2F3F460692B7083C60854A02"
 
+    def it_gets_all_items(self):
+        item_manager = ItemManager(self._fixture_path)
+        items = item_manager.get_all_items()
+
+        expected_item_uuids = [
+            '2E21D652E0754BD59F6B94B0323D0142',
+            '4A3D784D115F4279BDFCE46D0A162D57',
+            '5F7210FD2F3F460692B7083C60854A02',
+            '6371E49FEFA042EDB335421459E5B29F',
+            '9315F5EA8DCC4CB7BE09155DB7FCD1ED',
+            '97019BEBCF9E402F8F0C033474B1B85D',
+            '9E7673CCBB5B4AC9A7A8838835CB7E83',
+            'B851D6E3232842B0858BC10968632A9C',
+            'D05009E62D7D401CB8ACF2FE6981C031'
+        ]
+
+        assert len(items) == 9
+
+        for item in items:
+            assert item['uuid'] in expected_item_uuids
+
     def it_saves_items(self):
         self._init_default_data_dir()
         item_manager = ItemManager(self._temporary_path)
